# openapi-backend-stable.json
{"openapi":"3.1.0","info":{"title":"SkillMatch Pro","version":"0.1.0"},"paths":{"/auth/signup":{"post":{"tags":["Authentication"],"summary":"Signup","operationId":"signup_auth_signup_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["Authentication"],"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/me":{"get":{"tags":["Authentication"],"summary":"Read Users Me","operationId":"read_users_me_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/employees/":{"get":{"tags":["Employees"],"summary":"Read Employees","operationId":"read_employees_employees__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":20,"title":"Limit"}},{"name":"department","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Department"}},{"name":"position","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Position"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EmployeeOut"},"title":"Response Read Employees Employees  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Employees"],"summary":"Create Employee","operationId":"create_employee_employees__post","security":[{"OAuth2PasswordBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/employees/{employee_id}":{"get":{"tags":["Employees"],"summary":"Get Employee","operationId":"get_employee_employees__employee_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"employee_id","in":"path","required":true,"schema":{"type":"integer","title":"Employee Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Employees"],"summary":"Update Employee","operationId":"update_employee_employees__employee_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"employee_id","in":"path","required":true,"schema":{"type":"integer","title":"Employee Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Employees"],"summary":"Delete Employee","operationId":"delete_employee_employees__employee_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"employee_id","in":"path","required":true,"schema":{"type":"integer","title":"Employee Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/candidates/candidates/upload_cv":{"post":{"tags":["Candidates","Candidates"],"summary":"Upload Cv","operationId":"upload_cv_candidates_candidates_upload_cv_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_cv_candidates_candidates_upload_cv_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CandidateUploadRespose"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Internal Server Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/match/jobs":{"post":{"tags":["matching"],"summary":"Rank Candidates","operationId":"rank_candidates_match_jobs_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobMatchRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobMatchResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/jobs/":{"get":{"tags":["jobs"],"summary":"List Jobs","operationId":"list_jobs_jobs__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/JobOut"},"type":"array","title":"Response List Jobs Jobs  Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["jobs"],"summary":"Create Job","operationId":"create_job_jobs__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/jobs/{job_id}":{"get":{"tags":["jobs"],"summary":"Get Job","operationId":"get_job_jobs__job_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"integer","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["jobs"],"summary":"Update Job","operationId":"update_job_jobs__job_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"integer","title":"Job Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["jobs"],"summary":"Delete Job","operationId":"delete_job_jobs__job_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"integer","title":"Job Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/jobs/{job_id}/skills":{"get":{"tags":["jobs"],"summary":"List Job Skill Requirements","operationId":"list_job_skill_requirements_jobs__job_id__skills_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"integer","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/JobSkillRequirementOut"},"title":"Response List Job Skill Requirements Jobs  Job Id  Skills Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["jobs"],"summary":"Upsert Job Skill Requirement","operationId":"upsert_job_skill_requirement_jobs__job_id__skills_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"integer","title":"Job Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobSkillRequirementRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobSkillRequirementOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/jobs/{job_id}/skills/{skill_id}":{"delete":{"tags":["jobs"],"summary":"Remove Job Skill Requirement","operationId":"remove_job_skill_requirement_jobs__job_id__skills__skill_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"integer","title":"Job Id"}},{"name":"skill_id","in":"path","required":true,"schema":{"type":"integer","title":"Skill Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/skills/":{"get":{"tags":["skills"],"summary":"List Skills","operationId":"list_skills_skills__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/SkillOut"},"type":"array","title":"Response List Skills Skills  Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["skills"],"summary":"Create Skill","operationId":"create_skill_skills__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SkillCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SkillOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/skills/{skill_id}":{"get":{"tags":["skills"],"summary":"Get Skill","operationId":"get_skill_skills__skill_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"skill_id","in":"path","required":true,"schema":{"type":"integer","title":"Skill Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SkillOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["skills"],"summary":"Update Skill","operationId":"update_skill_skills__skill_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"skill_id","in":"path","required":true,"schema":{"type":"integer","title":"Skill Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SkillUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SkillOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["skills"],"summary":"Delete Skill","operationId":"delete_skill_skills__skill_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"skill_id","in":"path","required":true,"schema":{"type":"integer","title":"Skill Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/skills/employees/{employee_id}":{"post":{"tags":["skills"],"summary":"Assign Skill To Employee","operationId":"assign_skill_to_employee_skills_employees__employee_id__post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"employee_id","in":"path","required":true,"schema":{"type":"integer","title":"Employee Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeSkillAssignRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmployeeSkillOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/skills/employees/{employee_id}/{skill_id}":{"delete":{"tags":["skills"],"summary":"Unassign Skill From Employee","operationId":"unassign_skill_from_employee_skills_employees__employee_id___skill_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"employee_id","in":"path","required":true,"schema":{"type":"integer","title":"Employee Id"}},{"name":"skill_id","in":"path","required":true,"schema":{"type":"integer","title":"Skill Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/departements/":{"get":{"tags":["departements"],"summary":"List Departements","operationId":"list_departements_departements__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"search","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Search"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DepartementOut"},"title":"Response List Departements Departements  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["departements"],"summary":"Create Departement","operationId":"create_departement_departements__post","security":[{"OAuth2PasswordBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DepartementCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DepartementOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/departements/{departement_id}":{"get":{"tags":["departements"],"summary":"Get Departement","operationId":"get_departement_departements__departement_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"departement_id","in":"path","required":true,"schema":{"type":"integer","title":"Departement Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DepartementOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["departements"],"summary":"Update Departement","operationId":"update_departement_departements__departement_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"departement_id","in":"path","required":true,"schema":{"type":"integer","title":"Departement Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DepartementUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DepartementOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["departements"],"summary":"Delete Departement","operationId":"delete_departement_departements__departement_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"departement_id","in":"path","required":true,"schema":{"type":"integer","title":"Departement Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Health Check","operationId":"health_check__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_login_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","format":"password","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"format":"password","title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_auth_login_post"},"Body_upload_cv_candidates_candidates_upload_cv_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_cv_candidates_candidates_upload_cv_post"},"CandidateUploadRespose":{"properties":{"filename":{"type":"string","title":"Filename"},"skills":{"items":{"type":"string"},"type":"array","title":"Skills"},"preview":{"type":"string","title":"Preview"}},"type":"object","required":["filename","skills","preview"],"title":"CandidateUploadRespose"},"DepartementCreate":{"properties":{"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","required":["name"],"title":"DepartementCreate"},"DepartementOut":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["id"],"title":"DepartementOut"},"DepartementUpdate":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"}},"type":"object","title":"DepartementUpdate"},"EmployeeCreate":{"properties":{"employeeNumber":{"type":"string","title":"Employeenumber"},"first_name":{"type":"string","title":"First Name"},"last_name":{"type":"string","title":"Last Name"},"full_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Name"},"email":{"type":"string","format":"email","title":"Email"},"departement":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Departement"},"position":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Position"},"performance_score":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Performance Score"},"hire_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Hire Date"}},"type":"object","required":["employeeNumber","first_name","last_name","email"],"title":"EmployeeCreate"},"EmployeeOut":{"properties":{"employeeNumber":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Employeenumber"},"first_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"First Name"},"last_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Name"},"full_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Name"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"},"departement":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Departement"},"position":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Position"},"performance_score":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Performance Score"},"hire_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Hire Date"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["id"],"title":"EmployeeOut"},"EmployeeSkillAssignRequest":{"properties":{"skill_id":{"type":"integer","title":"Skill Id"},"level":{"type":"integer","maximum":5.0,"minimum":1.0,"title":"Level","default":1}},"type":"object","required":["skill_id"],"title":"EmployeeSkillAssignRequest"},"EmployeeSkillOut":{"properties":{"id":{"type":"integer","title":"Id"},"employee_id":{"type":"integer","title":"Employee Id"},"skill_id":{"type":"integer","title":"Skill Id"},"level":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Level"}},"type":"object","required":["id","employee_id","skill_id"],"title":"EmployeeSkillOut"},"EmployeeUpdate":{"properties":{"employeeNumber":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Employeenumber"},"first_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"First Name"},"last_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Last Name"},"full_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Name"},"email":{"anyOf":[{"type":"string","format":"email"},{"type":"null"}],"title":"Email"},"departement":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Departement"},"position":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Position"},"performance_score":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Performance Score"},"hire_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Hire Date"}},"type":"object","title":"EmployeeUpdate"},"ErrorDetail":{"properties":{"code":{"type":"string","title":"Code"},"message":{"type":"string","title":"Message"}},"type":"object","required":["code","message"],"title":"ErrorDetail"},"ErrorResponse":{"properties":{"detail":{"$ref":"#/components/schemas/ErrorDetail"}},"type":"object","required":["detail"],"title":"ErrorResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"JobCreate":{"properties":{"title":{"type":"string","title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"departement":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Departement"}},"type":"object","required":["title"],"title":"JobCreate"},"JobMatchRequest":{"properties":{"job_title":{"type":"string","title":"Job Title"},"required_skills":{"items":{"type":"string"},"type":"array","title":"Required Skills","default":[]},"min_experience":{"type":"integer","minimum":0.0,"title":"Min Experience","default":0},"limit":{"type":"integer","maximum":200.0,"minimum":1.0,"title":"Limit","default":20}},"type":"object","required":["job_title"],"title":"JobMatchRequest"},"JobMatchResponse":{"properties":{"job_title":{"type":"string","title":"Job Title"},"required_skills":{"items":{"type":"string"},"type":"array","title":"Required Skills"},"min_experience":{"type":"integer","title":"Min Experience"},"results":{"items":{"$ref":"#/components/schemas/MatchCandidateOut"},"type":"array","title":"Results"}},"type":"object","required":["job_title","required_skills","min_experience","results"],"title":"JobMatchResponse"},"JobOut":{"properties":{"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"departement":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Departement"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["id"],"title":"JobOut"},"JobSkillRequirementOut":{"properties":{"id":{"type":"integer","title":"Id"},"job_id":{"type":"integer","title":"Job Id"},"skill_id":{"type":"integer","title":"Skill Id"},"required_level":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Required Level"},"weight":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Weight"}},"type":"object","required":["id","job_id","skill_id"],"title":"JobSkillRequirementOut"},"JobSkillRequirementRequest":{"properties":{"skill_id":{"type":"integer","title":"Skill Id"},"required_level":{"type":"integer","maximum":5.0,"minimum":1.0,"title":"Required Level","default":3},"weight":{"type":"number","exclusiveMinimum":0.0,"title":"Weight","default":1.0}},"type":"object","required":["skill_id"],"title":"JobSkillRequirementRequest"},"JobUpdate":{"properties":{"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"departement":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Departement"}},"type":"object","title":"JobUpdate"},"MatchCandidateOut":{"properties":{"employee_id":{"type":"integer","title":"Employee Id"},"full_name":{"type":"string","title":"Full Name"},"score":{"type":"number","title":"Score"}},"type":"object","required":["employee_id","full_name","score"],"title":"MatchCandidateOut"},"SkillCreate":{"properties":{"name":{"type":"string","title":"Name"}},"type":"object","required":["name"],"title":"SkillCreate"},"SkillOut":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["id"],"title":"SkillOut"},"SkillUpdate":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"}},"type":"object","title":"SkillUpdate"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","token_type"],"title":"Token"},"UserCreate":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"},"role":{"anyOf":[{"type":"string","enum":["admin","user"]},{"type":"null"}],"title":"Role"}},"type":"object","required":["email","password"],"title":"UserCreate"},"UserResponse":{"properties":{"id":{"type":"integer","title":"Id"},"email":{"type":"string","format":"email","title":"Email"},"is_active":{"type":"boolean","title":"Is Active"},"role":{"type":"string","enum":["admin","user"],"title":"Role"}},"type":"object","required":["id","email","is_active","role"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"auth/login"}}}}}}